elderly_care_companion_backend/requirements.txt
Django
djangorestframework
psycopg2-binary
firebase-admin
celery
